CC-4575 [CC-3625] Schedule prices:Exception is occurred

+++ Bundles/PriceProduct/src/Spryker/Zed/PriceProduct/Business/PriceProduct/PriceProductRemover.php
@@ -65,10 +65,13 @@ class PriceProductRemover implements PriceProductRemoverInterface
 
             if ($idPriceProductStore !== null) {
                 $this->priceProductStoreWriterPluginExecutor->executePriceProductStorePreDeletePlugins($idPriceProductStore);
-                $this->priceProductEntityManager->deletePriceProductStore($idPriceProductStore);
+                $this->priceProductEntityManager->deletePriceProductDefaultById($priceProductTransfer->getPriceDimension()->getIdPriceProductDefault());
+                $this->priceProductEntityManager->deletePriceProductStoreByPriceProductTransfer($priceProductTransfer);
             }
 
-            $this->priceProductEntityManager->deletePriceProductById($priceProductTransfer->getIdPriceProduct());
+            if ($this->priceProductRepository->isPriceProductUsedForOtherCurrencyAndStore($priceProductTransfer) === false) {
+                $this->priceProductEntityManager->deletePriceProductById($priceProductTransfer->getIdPriceProduct());
+            }
         });
 
         $this->getLogger()->warning(sprintf('Price for product with id "%s" was deleted', $priceProductTransfer->getIdPriceProduct()));
